requirements.txt
torch>=2.5.0
torchvision>=0.20.0
torchaudio>=2.5.0
transformers>=4.46.0
tokenizers>=0.20.0
sentencepiece>=0.2.0
protobuf>=5.28.0
safetensors>=0.4.5
fasttext>=0.9.2
numpy==1.26.4
pandas==2.2.2
pyarrow>=18.0.0
polars>=1.6.0
tqdm>=4.67.0
scikit-learn==1.4.2
unicode-tr>=0.0.1
sentence-transformers
datasets
python-dotenv>=0.19.0
openai>=1.51.0
faiss-gpu>=1.7.4
joblib>=1.3.0
wandb

